requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
google-generativeai>=0.8.0
stripe==7.11.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
